pyproject.toml
[project]
name = "sanity-backup-to-s3"
version = "0.1.0"
description = ""
authors = [
    {name = "Weirdo",email = "adroszak@weirdo.codes"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3 (>=1.36.15,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "requests-toolbelt (>=1.0.0,<2.0.0)"
]


[tool.poetry]
packages = [
    { include = "src" }
]
[tool.poetry.dependencies]
python = "^3.12"

[tool.poetry.scripts]
sanity = "src.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
